+++ CHANGELOG.md
@@ -86,6 +86,7 @@
 * Ruffle OpenGL ES GUI fixed by ImGui update (#604)
 * Fix ImGui X11 to append selected events instead of replacing them
 * Scale lua draws when game window is detached
+* Scale pointer coordinates when game window is detached
 
 ## [1.4.5] - 2023-10-22
 ### Added